flask_with_jwt_extended/requirements.txt
Flask_JWT
Flask_RESTful
Flask-SQLAlchemy
Flask-JWT-Extended
jsonify